Automate deployment with Capistrano
This is a more conventional deployment mechanism and handles assets and
migrations, better handles failed deploys, and minimizes downtime. It
can also be kicked off from your local machine provided you have the
proper SSH config and the kerberos ticket.

I had to use a fork of the `gssapi` gem, which is what allows ruby to do
kerberos over `net/ssh` because of a segfault in the released version on
OS X. See: https://github.com/zenchild/gssapi/issues/12