Improve permissions handling on homepage
We were previously only receiving a list of departments from the
`RolesDb` and refering to this as `Roles`. In reality, we want the
permissions to be made up of departments and an associated access level
for each of those departments. This refactoring is a step in that
direction.

`Roles` have been replaced by `Permissions`, which model the concept of
department and access levels. We don't yet take advantage of the access
level, but this change will come later.

We now more efficiently assemble the list of authorized departments for
the homepage.

The previously existing cucumber scenarios were covered by a combination
of feature specs and controller specs. Feature specs have access to a
"grant_aceess" helper to facilitate faking out the RolesDB external
system interaction. Additionally, a "back door" was added for faking
touchstone logins from feature specs in the test environment.