Update design to use bourbon
By using bourbon, neat, bitters, and refills, we can get a reasonable
looking design without having to add a number of non-semantic classes.

This change loses the breadcrumb functionality of the previous layout,
but this can be added back in a more robust manner if desired. User
testing seemed to to show people weren't finding the breadcrumbs very
useful.

I also added headings to some pages and extracted translations. I have
focused on the home page, department page, and the listing of course
objectives thus far. We can move the designs over in this direction

We will need to update the design of other parts of the application as
we touch them for the task-based workflow changes we're making.