Introduce pundit for authorization
Previously, only the home page had any sort of authorization associated
to it. This change introduces Pundit so we can start authorizing all
actions based on the roles the user has the access required for the
action. I have not added unit tests for the policies as of yet, as the
sole policy is sufficiently tested via feature and controller specs.

This also introduces a `PermissionSet` object which allows us to ask
questions of the entire set of permissions in an easy way.