Rename department.role_department to slug
This field maps to the department name from the roles database. It is a
unique, string representation of the department. We typically refer to
these as slugs. `slug` is less redundant, shorter, and I believe less
confusing than `role_department`.

I have also added a unique index to ensure that we don't duplicate these
values and make them slightly quicker to search for.