Add outcome management dashboard
This is the entry point for the outcome management (admin) workflow.
It's focused on letting the user know what they might have to do next
and a little bit of the why (content to be edited by Jon later).

To facilitate listing courses that do not have outcomes, I added a
counter cache to courses so they always know how many outcomes are
associated.

Determining courses that have custom outcomes that are not fully aligned
is handled via N+1 query at the moment. This can be addressed in the
future, but most people will have access only to a few courses so it's
not the biggest issue presently.