Enable creation of direct assessment starting with subject
https://trello.com/c/ymayP0kT

Instead of having a user choose if they want to create an assessment "by
course" or "by subject", we now take users through two different
workflows when adding direct and indirect assessments.

Indirect assessment creation is similar to the previous workflow; users
choose a course and an outcome and add an assessment.

When users create new direct assessments, they are prompted first to
choose a subject from a dropdown, then are shown all outcomes for the
departments they have access to with check boxes, so they can pick the
ones they want to associate with the new assessment.

Since we display all outcomes associated with departments the user has
access too,  we added a validation to ensure that the outcomes chosen
belong to the same department.