Add cancel buttons to forms
This avoids UI dead ends on the forms, which are the leaf nodes of our
workflows. The cancel button is equivalent to the back button and is
accomplished with JavaScript. On the off chance the user does not have
JavaScript enabled, the cancel button will not appear at all (because it
would be nonfunctional if it did).