Add validations to outcomes form
This adds missing database constraints and indexes on outcomes and then
backs them up with validations in the form. This leads to a better user
experience - they have a visual indication of required fields and get
nice errors rather than server errors when things fail.

As part of this change, I configured simple_form to not display error
text for each individual field, instead opting for a summary at the top
of the form. This isn't ideal; Inline error messages would be best with
the summary being reserved for things that relate to more than one
field. However, this was simpler to style in a reasonable manner for
now.