Add back ability to archive from results
This now uses a dedicated controller and action and a feature test. The
use of `default_scope` replicated previous functionality but may not be
the long term solution. Default scope tends to leak places where you
would really rather it didn't. For this small use case right now, where
the default scope matches the boolean default for a new record, it's not
an immediate issue.