Remove default scope on direct assessments
This proved to be a problem when trying to create a UI for unarchiving
assessments. You can call `unscoped` to remove all previous scopes
(including the default scope) but that means your scopes are more order
dependent than desired.

This change updates subject so it is related only to unarchived direct
assessments. There's no context in the app where we'd care to know about
a subjects archived direct assessments.

This also means we now have an interface for unarchiving assessments
because archived assessments will once again show up on the "manage
assessments" screen for a given course. This is probably not the final
UI here, but this is a step in the correct direction.