Track number of active assessments for an outcome
We were previously tracking the number of assessments associated to an
outcome with a counter cache on `OutcomeAssessment`. With the
introduction of the notion of archived assessments, this count no longer
reflects the number of active assessments.

Getting counter caches to obey conditions is tricky. Since the condition
we care about is binary it's not too bad to do it manually. We keep
`assessments_count` counting the overall number of assessments and add a
service object (AssessmentArchivist) that is responsible for archiving
an unarchiving that tracks a separate count of the number of archived
outcomes. Then it's simple math to get the number of active outcomes.