Rough in Gradebook Histogram
We can assist users in reporting results on a direct assessment by
allowing them to view the gradebook grade distribution histogram from
the gradebook service. This will be particularly helpful to users who do
not have access to the gradebook (because it provides personal student
information) but can have access to aggregate grade data for the
purposes of completing ABET assessments.

This change works off of local fixtures rather than integrating with the
actual gradebook service. That will be in the next change.