Add adapters to GradebookClient
This will let us switch between a version that serves fixtures and a
version that actually makes requests. It gives me a place to make that
code diverge.

For rails auto loading reasons, I had to rename the `Gradebook`
namespace to `Gradebooks` because rails was confused by the `module
Gradebook` in the controller and the `class Gradebook` inside the
`GradebookClient` module. There's possibly a better solution to this
than the rename, but I couldn't immediately find it.