Allow users to filter assessments by outcome
This adds a form to the assessments index that lets users filter to a
specific outcome.

With the ability to filter by outcome, the need to display the outome
associated to each assessment is lessened. We can now show uniq direct
and indirect assessments in their respective tables without the banner
row.