Implement real gradebook service adapter
We're now able to use the `LearningModules` adapter to connect to the
actual gradebook service.

If you are running in production, you will *always* try to use the
actual adapter.

If you are running in development mode and have the client cert on your
machine, you will use the actual adapter. In all other environments, you
default to using the fake adapter (which returns fixtures for all
inputs).

This change also adds the ability to switch the number of bins in the
histogram.