Add result activity feed
Users wanted to quickly be able to see if results were being added for
the departments they have access to.

The result activity feed is built off of data that PaperTrail already
has for results. It tracks every create, update, and delete so we don't
have to. I created an `Activity` model that has some additional smarts
about what we want to get out PaperTrail data (the associated user, the
associated, non-polymorphic result).

To accommodate authorization in a performant way, I denormalized the
department a result is associated to. Previously the result department
came from `assessment -> outcomes -> courses -> departments`. Due to the
polymorphic assessment relationship, we weren't going to be able to
efficiently (read: In SQL) limit the results to only those in
departments the user can access. By caching the department at save
time, we can quickly get at this data.

One downside of this is that we can't easily show activities for deleted
results. The result is no longer in the results table, so we can't tell
what department it belongs to. We may be able to rehydrate this data
from the PaperTrail version, but doing so would be non-trivial. Earlier
John had mentioned that maintaing deletes was not a priority, so I went
with this for now.