Use fake roles on staging for now
There's some more debugging that has to happen with the RoleDB on the
staging server. For now, we can have it use the fake roles that we use
in development. This will allow Jon to review things on staging (as his
user is associated to all the fake permissions. Other users will still
have no department access.

This is meant to let us move forward with acceptance on staging.