Add API to get outcomes for a subject
Faculty want to be able to display outcomes associated with a
class/subject on their own websites. This API will allow them to do
that.

The API format is extremely simplistic at this point. Jon is checking to
see if there are preferred standards (JSON API?) within use at MIT. In
the meantime, I thought this simplified format would likely be of the
most use to the people we think will be consuming this API.