Show list of accessible reports if necessary
Some users have access to more than one course that has reports. In this
case, we are supposed to show them a list of accessible reports, but due
to a bug we were redirecting to their first accessible course instead.
We only want to redirect if they have access to exactly one course.