Remove indirect assessments
MIT is not making use of the indirect assessment feature and continuing
to have it in the interface leads to confusion. This change removes all
references to indirect assessments. This makes "direct assessment"
unnecessarily specific (just 'assessment' would likely do just fine) but
we can leave that change for a future PR.