Add a custom outcome button appears every course
When the user clicks the "Adopt Standard Outcomes" button from the
"Manage Outcomes Dashboard" and then returns to the dashboard and clicks
on "View Outcomes", an "Add a Custom Outcome" button should appear
regardless of whether the course has unaligned or fully aligned
outcomes.

This branch contains partially refactored tests.