Add outcome coverage
This branch adds outcome coverage to a subject/class. Outcome
coverage refers to the one or more outcomes covered by any given
subject/class in a course. Connecting outcomes to subjects/classes and
subjects/classes to courses means that the UI can show a user which
outcomes have been matches to a class and which outcomes still need to
be demonstrated by a subject/class in the course.

Outcome coverage belongs to all of the following: outcome, course, and
subject.