Add `Label` type to ensure uppercase labels
This types coerces any label values to uppercase both in queries and in
inserts. For example, `StandardOutcome.where(label: 'a')` will search
forstandard outcomes with a label of `A`. `Outcome.create(label: 'z')`
will similarly create an outcome with a label of `Z`.

This allows us to normalize the case at the app level.