Validate outcome and subject on outcome coverage
Users could previously create a coverage without any of the required
association. This updates the data model to require the associations and
then updates the model to add appropriate user-facing validations.

Validations were not added to the `course` association because it is not
user-facing. If we somehow attempt to create a coverage without a
course, this is a programming error and should result in an application
error.