Enable multiple outcomes per coverage
This allows a given class to say that it covers multiple outcomes. We do
this by adding an `OutcomeCoverage` association to `Coverage`, removing
the previously existing `belongs_to :outcome` there.

In this change, I updated the controllers, models, and tests to allow
for the possibility of multiple outcomes, but the UI does not yet expose
the ability to add a second outcome. That will come in a future change.