Test adding coverage via chromedriver
We will be adding some JavaScript to this page in order to support
adding many outcomes to a coverage dynamically. To get adequate test
coverage on this, we need to have a JavaScript test driver.

We usually use capybara-webkit for this, but QT was giving me troubles
and the world seems to be moving on to chromedriver. I decided to give
chromedriver a try here. Installation was an absolute picnic compared to
QT and capybara-webkit.

With the test case now running in a JavaScript driver, we can no longer
faithfully interact with the select boxes, which are actually hidden by
Selectize. I updated the test to interact with the Selectize input
instead.