Update to Rails 5.1
Rails 5.1 includes some features we may want to use to ease deployment -
namely the ability to check in an encrypted secrets file. It's also good
practice to stay up-to-date when possible.

This upgrade exposed a couple of issues where we are chaining `distinct`
on to chained relationships. Because those relationships have default
`ORDER BY` clauses the SQL `DISTINCT` fails because we're not
selecting/distincting over the things we are trying to order by. It's
unclear how or if this was working in Rails 5.0.

Long term, the solution is to remove default ordering from the
associations. It has many of the same disadvantages of default scope in
that it's difficult to override. Short term, the fix in this case was to
force the de-deuplication to occur in memory.