Introduce encrypted secrets
`secrets.yml.enc` is meant to contain encrypted secrets (the file
currently has no secrets in it). This will be useful for storing API
keys and the like.

Decrypting the secrets requires `config/secrets.yml.key`, which was
generated when I created the file, but will not be checked in to the
repository. I will share this key with Jon and editing production or
staging secrets will require access to this key.

We will also need to add the key to the environment on our production
and staging app servers.