Pre-populate coverages form with unmatched outcome
This branch allows users on the `manage_assessments/courses/:course_id`
page to add a single additional outcome to the listed course from this
page. When a user clicks on an outcome, they are redirected to the
`new_manage_assessments_course_coverages_path` with the selected outcome
already pre-populated. To accomplish this, the `outcome_id` is passed as
a query parameter in the url to the `Coverages` controller.