Allow adding attachments to coverages
When creating a coverage the user can optionally upload one or more
attachments. This UI is not what is reflected in the prototype, but the
functionality is in place. We can refine the UX in further iteration.

In dev and test, we use the file system for storage but use S3 for
staging and production. The S3 secret key is stored in encrypted secrets
storage.

MIT wants files uploaded here to remain private except to those with
appropriate access. For that reason, the files are uploaded to s3 with
private permissions. The URL for the file is set to a local path where
we will eventually implement our own controller that performs
authorization and redirects with an expiring URL to the actual file.
That part will be implemented in a future PR.