Refactor split tabs
The previous split-tab JavaScript was heavily dependent on the structure
of the markup. This implementation is instead dependent on being wired
up via data attributes.

The tab controls themselves must have a `data-tab` attribute set to some
identifier. The associated content must have a `data-content` with the
value set to the same value.

After wiring up the event handlers on the tabs, we trigger a click on
the first tab to set it as the active tab.

I also updated the existing test case for viewing unmatched outcomes to
use the JavaScript driver so we get coverage on this bit of JavaScript.