User can edit coverage for a course
A user can now add an outcome to an existing coverage or change an
existing outcome to a different outcome. To accomplish this
functionality, `edit` and `update` were added to the `Coverages`
controller. A user cannot yet delete an outcome from a coverage. If a
user tries to add a duplicate outcome to a coverage, she will be
prompted to select another outcome. User friendly error handling still
needs to be added to the `edit` form.