Use hidden file inputs for coverage attachments
The design calls for an "Add an attachment" link that prompts the user
for a file to upload rather than a standard file input. We accomplish
this by hooking in to the cocoon lifecycle events to trigger a click to
the now-hidden file input when we add a new association record via
cocoon.

We render the names of the files to be uploaded in a list but keep the
inputs hidden. This change also adds the ability to remove an
attachment.

I was unable to get the JavaScript test drivers to play nice with an
auto-triggered file input dialog backed by a hidden input. I originally
added a workaround that made the JavaScript behave differently when
running in test, but this didn't feel like a useful distinction. After a
good deal of time trying to get file attachments running appropriately
in this design, I abandoned the idea of testing them. I have manually
tested file uploads. I wish there were an easier way of testing this,
but I can't figure it out. I may come back to this if there's more time.