User can archive an outcome coverage
When a user clicks the Delete button on an outcome coverage, the
selected outcome coverage is archived. If the coverage that this now
archived outcome coverage belonged, no longer has any outcome coverages,
the coverage is also archived. To accomplish this functionality, an
`archived` boolean column was added to both the `Outcome Coverages` and
`Coverages` tables. An `Archive` service object handles the archiving of
outcome_coverages and associated coverages.