Report results on assignments
We want to record results against assignments rather than assessments.
Assessments are being replaced by assignments. This change updates the
models to accept results on assignments, but it leaves the old fields
and relationships in place. We still need to migrate existing data and
then we should be able to clean up old data and set, for example,
`assignment_id` on `results` to `null: false`.