Make recent migrations work against production
We had not been running these migrations against production data. Doing
so highlighted a couple of issues. The biggest was that we have some
`outcome_assessment` records that reference now-deleted
`IndirectAssessment` records. We checked with Jon, who said it was okay
to delete these. In this case, it's okay to edit past migrations because
they have not been deployed and run yet.

The `create_table` statements in `schema.rb` also changed to indicate
that the `id` column is of type `:serial`. This appears to be
ActiveRecord 5.1 being more explicit than it had been previously. The
underlying table definition is unchanged. See:
https://github.com/rails/rails/blob/c6b4b4a52c7ec66d1923ef44dcbc71b582d7b165/activerecord/lib/active_record/connection_adapters/postgresql/schema_definitions.rb#L49-L53