Correct `schema.rb`
The production database had marked the `AddMissingForeginKeys` migration
as run, but the index and foreign key it adds were not present. When I
copied the production database locally for testing, I ended up checking
in a `schema.rb` file that was missing these. To remedy this, we
manually ran the migration in all environments (staging and production)
and are now correcting the checked-in `schema.rb` file.