Populate subjects from the subject listings API
We were previously populating subjects from an in-repo CSV, which we
imported with a rake task. MIT offers the [subject listings API][1] as
the official source for this information.

This PR introduces a client for this API and converts the existing rake
task import to use the client as the source of data. The client defaults
to using a "fake" adapter which loads from the json fixture in the app.
If the app is configured with access keys for the API, the real API will
be used.

[1]:https://developer.mit.edu/apis-subjectlistings