Don't create duplicate assignments when migrating
The previous version of the assessment migration task was creating
duplicate assignments because it was creating the same assignment for
every outcome on the assessment.

Fixing this revealed an incompatibility in the data model between
assessments and coverages/assignments. With assessments, users could
create multiple assessments in the same subject that covered the same
outcome. With the change to coverages and assignments, we ask the users
to provide the assignment that *best* covers an outcome in the given
subject. We have validations and database constraints that prevent
covering the same outcome more than once in a subject.

As a result, when migrating assessments to outcome coverages, we mark
the outcome coverage associated with all but the most recently active
(most results, most recently created) assessment as archived.