Remove assessment-based associations from outcomes
The UI has been updated to use assignment and coverage associations
where we were previously relying on assessments. The assessment ->
assignment migration does not need any of these associations on
`Outcome` so they are safe to delete.

The `assessments` association had a counter cache defined, but this is
no longer necessary. We removed the corresponding test for this.