Don't use instance variables in partials
These can sneak in when we extract partials, but we should instead be
relying on local variables. Partials that use local variables are easier
to reuse as they don't rely on pre-existing instance state. They're also
easier to maintain as changing the name of an instance variable in a
controller doesn't cause downstream partials to break unnecessarily.