Add additional activities to the Activity Feed
This branch add PaperTrail to assignment and outcome coverage and
displays assignments and outcome coverages in the activity feed along
with results.

Clicking the link associated with creating or updating of a  result
directs the user to the `manage_results_assignment_path`, where the user
can view and edit the result. Clicking the link associated with creating
or updating assignment directs the user to the corresponding course page
(`manage_assessments_course_path`) for that assignment.