Sort outcomes alphabetically within each class
I tried and failed to do this within our `includes` statements in the
controller. It requires adding explicit associations that have ordering
provided, but getting those to work with our `has_many: :through`
relationships was proving difficult. We're not dealing with a ton of
records here, so an in memory sort shouldn't be bad.