Remove content type validation on attachments
We discussed this with Jon a couple of weeks ago who said it was
difficult to predict what types of files people might upload. Images,
MS Word docs, Excel files, Open Office, Text, PDF, PostScript... its all
a possibility.

We discussed the risks and Jon decided to opt not to restrict the
content type given that the set of users is restricted.