Activity feed links auto scroll
The activity feed reports when an assignment is created or updated,
and includes a link to the specific assignment being referenced. Now
when a user clicks on an assignment link, she will be directed to the
point on the courses page that lists that assignment. To accomplish this
feature, an anchor is passed to the `manage_assessments_course_path`
when the activity type is assignment. The `id` of that assignment is
added to the html on the courses page.