Order coverages by their subject number
When we display coverage by course, we want to order the coverages by
their subject. Currently, the order is non-deterministic.

This performs the sort in memory to avoid having to add an explicit
"ordered_coverages" association to Course to do it in SQL with eager
loading. I considered that option, but it would have also required
rewriting some other methods on `CourseCoverage` to assume use of the
ordered collection as well, and that didn't seem desirable.