Always show remove attachment link
If the user cancels a file attachment dialog, they previously had no way
to remove the empty file input. It wasn't necessary to do so, but I can
definitely see confusion resulting from that.

Ideally there would be an event we could handle when the user cancels
that dialog, but there is not. No good workarounds exist for this
either. As a stopgap solution, this change removes the code that was
keeping the "delete" link hidden until there was a file added.

Now the user can at least manually clean up after themselves.