Render all attachments above the "add" button
Previous to this change, pre-existing attachments would render above the
"Add and Attachment" button while new attachments would render below.
They now all render above the button, which is a sibling list element to
the attachments themselves.

I also updated the `fields_for` call so it doesn't render a hidden
`input` between the `li` tags for the attachment id, which is not valid
HTML. Instead, we manually render the `id` inside the `li`.