Display username on activity feed
This branch displays the username of the person responsible for the
activity listed on the Activity Feed. To accomplish this feature,
`before_action :set_paper_trail_whodunnit` has been added to the
`Applications Controller`. As of PaperTrail 5.0, this line is no longer
a default of PaperTrail.

PaperTrail change logs located here:
https://github.com/airblade/paper_trail/blob/master/CHANGELOG.md