Add outcome information to assignment form
The assignment form didn't tell the user what subject or outcome the
assignment was for.

In order to get this to work properly for new, unpersisted assignments,
I had to change the multi-level `has_one :through` associatins into
delegations. Rails isn't able to properly wire up the multi-level
associations in memory, but handling this via delegation works just
fine.